# documentation/api/identity.openapi.yaml
openapi: 3.1.0
info:
  title: Zagrosi Identity API
  version: 0.1.0
  summary: Public identity, SSO, SCIM, and token-management API.
  description: >
    This specification documents the browser auth, OIDC, SAML, SCIM,
    session, personal access token, SCIM token, service token, and admin
    unlock surfaces shipped by the identity foundation.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
...
servers:
  - url: https://api.zagrosi.example
    description: Production API gateway
security: []
...
paths:
...
  /v1/auth/sign-in:
    post:
      tags: [Auth]
      operationId: signIn
      summary: Verify email and password and issue browser session cookies.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        "200":
          description: Signed in.
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SetCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignInResponse"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /v1/api-tokens:
    get:
      tags: [Personal access tokens]
      operationId: listApiTokens
      summary: List the caller's personal access tokens.
      security:
        - SessionCookie: []
        - BearerToken: []
      responses:
        "200":
          description: Token list.
          content:
            application/json:
              schema:
                type: object
                required: [tokens]
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiTokenView"
        "403":
          $ref: "#/components/responses/Problem403"
    post:
      tags: [Personal access tokens]
      operationId: createApiToken
      summary: Mint a personal access token. The raw `pat_` value is returned once.
      security:
        - SessionCookie: []
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiTokenRequest"
      responses:
        "201":
          description: Token minted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedApiToken"
        "403":
          $ref: "#/components/responses/Problem403"
        "409":
          $ref: "#/components/responses/Problem409"
  /v1/api-tokens/{id}:
    get:
      tags: [Personal access tokens]
      operationId: getApiToken
      summary: Read personal access token metadata.
      security:
        - SessionCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Token metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTokenView"
        "404":
          $ref: "#/components/responses/Problem404"
    delete:
      tags: [Personal access tokens]
      operationId: deleteApiToken
      summary: Revoke a personal access token.
      security:
        - SessionCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: Token revoked.
        "404":
          $ref: "#/components/responses/Problem404"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_sid
      description: Secure HttpOnly SameSite=Lax Path=/ session cookie.
    CsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_csrf
      description: Double-submit CSRF cookie paired with `X-Zagrosi-CSRF`.
    OidcCsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_oidc_csrf
      description: OIDC callback state cookie.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: "pat_<43>, scim_<43>, svc_<43>"
    ScimBearer:
      type: http
      scheme: bearer
      bearerFormat: scim_<43>
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    SetCookie:
      schema:
        type: string
      description: >
        Session cookies use `__Host-zagrosi_sid; Path=/; Secure; HttpOnly;
        SameSite=Lax`. CSRF cookies use `__Host-zagrosi_csrf; Path=/; Secure;
        SameSite=Lax`.
    RetryAfter:
      schema:
        type: integer
        minimum: 0
    RateLimitLimit:
      schema:
        type: integer
    RateLimitRemaining:
      schema:
        type: integer
    RateLimitReset:
      schema:
        type: integer
    WWWAuthenticate:
      schema:
        type: string
      examples:
        bearer:
          value: Bearer
        scim:
          value: Bearer realm="SCIM"
...
  responses:
...
    Problem400:
      description: Bad request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Problem401:
      description: Unauthorized.
      headers:
        WWW-Authenticate:
          $ref: "#/components/headers/WWWAuthenticate"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Problem403:
      description: Forbidden.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Problem404:
      description: Not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Problem409:
      description: Conflict.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    RateLimited:
      description: Rate limited.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    SignInRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SignInResponse:
      type: object
      required: [session]
      properties:
        session:
          $ref: "#/components/schemas/Session"
...
    Session:
      type: object
      required: [id, user_id, org_id, expires_at, version, amr]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
        amr:
          type: array
          items:
            type: string
        acr:
          type:
            - string
            - "null"
...
    CreateApiTokenRequest:
      type: object
      required: [display_name, scopes]
      properties:
        display_name:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
    ApiTokenView:
      type: object
      required: [id, display_name, scopes, created_at]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        scopes:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        revoked_at:
          type:
            - string
            - "null"
          format: date-time
    IssuedApiToken:
      allOf:
        - $ref: "#/components/schemas/ApiTokenView"
        - type: object
          required: [token]
          properties:
            token:
              $ref: "#/components/schemas/PrefixedToken"
...
    PrefixedToken:
      type: string
      pattern: "^(sid|pat|scim|svc|vrf|rst)_[A-Za-z0-9_-]{43}$"
      description: Token prefix is part of the SHA-256 hash input.
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string